feat: 模型字段优化与数据库会话上下文管理器重构
Boolean 字段替换 Integer (is_idle/is_active/is_enabled 等)，ChatSession/Message 新增 model_provider 字段，FTS5 支持增量维护；get_background_db_session 改为上下文管理器模式简化资源释放；任务领取改为原子 UPDATE+子查询避免并发竞争；Electron 请求错误改为 reject 而非静默兜底；前端提取 formatTime/formatDuration 至 utils/format，浏览器 dev 模式支持 localStorage 存储 API Key。

+++ backend/app/models/audit.py
@@ -4,7 +4,7 @@ from __future__ import annotations
 
 from datetime import datetime
 
-from sqlalchemy import Column, DateTime, Integer, String, Text, func
+from sqlalchemy import BigInteger, Column, DateTime, Integer, String, Text, func
 
 from app.core.database import Base
 
@@ -17,7 +17,7 @@ class AuditLog(Base):
     id = Column(Integer, primary_key=True, autoincrement=True, comment="主键 ID")
     action = Column(String(64), nullable=False, index=True, comment="操作类型，如 'memory.delete', 'model.config.save'")
     target_type = Column(String(64), nullable=True, index=True, comment="操作对象类型，如 'memory', 'session'")
-    target_id = Column(Integer, nullable=True, comment="操作对象 ID")
+    target_id = Column(BigInteger, nullable=True, comment="操作对象 ID")
     summary = Column(String(256), nullable=True, comment="操作摘要（不含敏感正文）")
     detail = Column(Text, nullable=True, comment="操作补充信息（JSON 字符串，不含敏感正文）")
     result = Column(Integer, nullable=False, default=0, comment="操作结果，0=成功，1=失败")
